added Edit Schedule views
