+++ views/doctor/doctor.html
@@ -179,12 +179,12 @@
                       <input type="text" class="form-control" id="specialization" required />
                     </div>
                     
-                    <div class="card-body">
-                      <button class="w3-bar-item w3-button w3-red" type="" onclick="addForm(event)">Add Working Hours +  </button>
+                    
+                      
+                        
+                      <button style="margin:10px;" onclick="addForm()" class="mx-1 btn btn-light" id="addButton" type="button" name="addButton" >Add Working Hours + </button>
 
                       <div class="container" id="container1">
-                        
-                      
                       
                     </div>
