+++ views/doctor/editDocSchedule.js
@@ -1,74 +1,409 @@
-/**
- another form window to edit/show doctor schedule
- **/
-const path = require("path");
-const {
-  BrowserWindow,
-  ipcMain
-} = require("electron");
-const {
-  updateUser,
-  getAllUsers,
-} = require("../models/user.js");
+// DOM Nodes
+const cancelButton = document.getElementById('dismiss-window');
+const deleteButton = document.getElementById('delete-doctor-sch');
+const editButton = document.getElementById('edit-doctor-sch');
+const errorDiv = document.getElementById('error');
+let serverUrl
+let workCount = 0;
 
-const { removeEventListeners } = require("../ipcHelper.js");
+// clean up
+window.onUnload = () => window.editContentAPI.removeListeners();
 
 
-let win
+window.editContentAPI.receive('response-doctor-data', async data => {
+  try {
+
+    serverUrl = localStorage.getItem("serverUrl");
+    if (!serverUrl || serverUrl === null)
+      throw new Error ("Erorr: failed to get server url");
+
+    await showDoctor(data._id);
+
+    toggleInputs(data.method);
+  }
+  catch (error) {
+
+  }
+});
+
+// dismiss/close form window
+cancelButton.addEventListener('click', () => {
+  window.editContentAPI.send('dismiss-form-window', '');
+})
 
 
-exports.createDocEditWindow = function createDocEditWindow(parentWindow, type, contents) {
+// edit/update doctor
+editButton.addEventListener('click', async e => {
 
-  if (!win || win === null) {
-    win = new BrowserWindow ({
-      width: 550,
-      height: 700,
-      parent: parentWindow,
-      modal: true,
-      show: false,
-      backgroundColor: '#ffffff',
-      webPreferences: {
-        contextIsolation: true,
-        nodeIntegration: false,
-        preload: path.join(__dirname, "../preload_scripts/editFormPreload.js")
+  e.preventDefault();
+  e.target.setAttribute("disabled", true);
+  e.target.innerHTML = "Loading ...";
+
+  try {
+
+    const id = document.getElementById('id')?.value;
+    const dId = document.getElementById('doctorId')?.value;
+    const name = document.getElementById('name')?.value;
+    const workingSch = document.getElementById('workingSchedule')?.value;
+
+
+    //const specialization = document.getElementById('specialization')?.value;
+    
+
+    if (!id || id === '' ||!dId || dId === '' ||!name || name === '' ||!workingSch || workingSch === '' ) {
+      throw new Error ("Missing Required Inputs");
+    }
+
+    const response = await editDoctorById(id, {
+      name,
+      specialization,
+    
+    });
+    console.log(response);
+    if (response && response.ok) {
+      // update opreration successful
+      // inform the main process that new data update is done
+      console.log(await response.json());
+      window.editContentAPI.send('doctor-form-finish');
+    }
+    else {
+      const { message } = await response.json();
+      const error = message ? message : "Error editing doctor. code: 500";
+      showErrorMessage(error);
+    }
+  }
+  catch(error) {
+    console.log('Error Fetching Update Doctor Response', error);
+    showErrorMessage(`Application Error: code 300`);
+  }
+  finally {
+    e.target.removeAttribute("disabled");
+    e.target.innerHTML = "Edit";
+  }
+});
+
+
+deleteButton.addEventListener("click", async e => {
+  try {
+    e.target.setAttribute("disabled", true);
+    e.target.innerHTML = "Loading ...";
+
+    const id = document.getElementById("id")?.value;
+
+    const response = await deleteDoctorById (id);
+
+    if (response && response.ok) {
+      window.editContentAPI.send('doctor-form-finish');
+    }
+    else {
+      const { message } = await response.json();
+      const errorMessage = message ? message : "Error: deleting doctor. code 500";
+      showErrorMessage(errorMessage);
+    }
+  }
+  catch (error) {
+    showErrorMessage(`Application Error: code 300`);
+  }
+  finally {
+    e.target.removeAttribute("disabled");
+    e.target.innerHTML = "Delete";
+  }
+});
+
+
+async function showDoctor (id) {
+  try {
+    const response = await getDoctorById(id);
+
+    if (response && response.ok) {
+      const emp = await response.json();
+      //console.log(emp);
+      displayDoctorData (emp);
+    }
+    else {
+      const { message } = await response.json();
+      const errorMessage = message ? message : "Error: failed to get doctor data. code 500";
+      showErrorMessage(errorMessage);
+    }
+  }
+  catch (error) {
+    console.error(error);
+    //showErrorMessage(`Application Error: code 300`);
+  }
+}
+
+/*  Adding new inputs for working hours */
+async function addForm(event){
+  workCount+= 1;
+  const container = document.getElementById("newContainer");
+  console.log(container.childElementCount);
+
+  var row1 = document.createElement("div");
+    row1.id = "row1";
+    row1.class = "row";
+
+    var values = ["Monday","Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"];
+ 
+    var select = document.createElement("select");
+    select.name = "days";
+    select.id = "days" +workCount;
+ 
+    for (const val of values)
+    {
+        var option = document.createElement("option");
+        option.value = values.indexOf(val);
+        option.text = val.charAt(0).toUpperCase() + val.slice(1);
+        select.appendChild(option);
+    }
+ 
+    var label = document.createElement("label");
+    label.innerHTML = "Working Days: "
+    label.htmlFor = "days";
+    label.id = "wLabel";
+    
+ 
+    row1.appendChild(label).appendChild(select);
+
+
+    row1.appendChild(document.createTextNode("Start Time"));
+    let input1 = document.createElement("input");
+    input1.type = "time";
+    input1.name = "startTime";
+    input1.id = "startTime" + workCount;
+    
+    row1.appendChild(input1);
+
+
+    row1.appendChild(document.createTextNode("End Time"));
+    let input2 = document.createElement("input");
+    input2.type = "time";
+    input2.name = "endTime";
+    input2.id = "endTime" + workCount;
+    
+    row1.appendChild(input2);
+
+    let removeBtn = document.createElement('button');
+    removeBtn.setAttribute('class', 'w3-bar-item w3-button w3-red');
+    removeBtn.innerHTML = 'Remove';
+    row1.appendChild(removeBtn);
+    container.appendChild(row1);
+
+
+    removeBtn.addEventListener('click', e => {
+      for(i=0; i< 6; i++){
+        row1.removeChild(container.lastChild);
       }
+      console.log(container.childElementCount);
+      workCount--;
+    
     });
 
-
-    win.loadFile(path.join(__dirname, "../views/user/editDoctor.html"));
-    //win.openDevTools();
+}
 
 
-    win.once("ready-to-show", () => win.show());
 
-    win.on("close", () => {
-      if(win) {
-        removeEventListeners(ipcMain, ["dismiss-form-window", "from-data-finish","doctor-form-finish"]);
-        removeEventListeners(win.webContents, ["did-finish-load"]);
-        win = null;
+function displayDoctorData(emp) {
+
+  const id = document.getElementById('id');
+  const dId = document.getElementById('doctorId');
+  const name = document.getElementById('name');
+  const container = document.getElementById('currentContainer');
+
+  const { fullname } = emp;
+
+  id.value = emp._id;
+  dId.value = emp.doctorId;
+  name.value = emp.name;
+  const workingSch = emp.workingSchedule;
+  console.log(workingSch.length);
+  for (var i = 0; i < workingSch.length ; i++)
+  {
+    var row1 = document.createElement("div");
+    row1.id = "row2";
+    row1.class = "row";
+
+
+    var values = ["Monday","Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday"];
+ 
+    var select = document.createElement("select");
+    select.name = "days";
+    select.id = "days" +workCount;
+ 
+    for (const val of values)
+    {
+        var option = document.createElement("option");
+        option.value = values.indexOf(val);
+        option.text = val.charAt(0).toUpperCase() + val.slice(1);
+        select.appendChild(option);
+    }
+ 
+    var label = document.createElement("label");
+    label.innerHTML = "Working Days: "
+    label.htmlFor = "days";
+    label.id = "wLabel";
+    select.value = workingSch[i].day;
+  
+ 
+    row1.appendChild(label).appendChild(select);
+
+
+    row1.appendChild(document.createTextNode("Start Time"));
+    let input1 = document.createElement("input");
+    input1.type = "time";
+    input1.name = "startTime";
+    input1.id = "startTime" + workCount;
+    input1.value = workingSch[i].startTime;
+    row1.appendChild(input1);
+
+
+    row1.appendChild(document.createTextNode("End Time"));
+    let input2 = document.createElement("input");
+    input2.type = "time";
+    input2.name = "endTime";
+    input2.value = workingSch[i].endTime;
+    input2.id = "endTime" + workCount;
+
+    row1.appendChild(input2);
+
+    let removeBtn = document.createElement('button');
+    removeBtn.setAttribute('class', 'w3-bar-item w3-button w3-red');
+    removeBtn.innerHTML = 'Remove';
+    row1.appendChild(removeBtn);
+    container.appendChild(row1);
+
+
+    removeBtn.addEventListener('click', e => {
+      for(i=0; i< 6; i++){
+        row1.removeChild(row1.lastChild);
       }
-    })
-
-    win.webContents.on("did-finish-load", () => {
-      win.webContents.send("response-doctor-data", {_id: contents, method: type});
-    });
-
-
-    /**
-    IPC Messages
-    **/
-
-    /* Dimiss Window */
-    ipcMain.on("dismiss-form-window", (event, args) => {
-      if(win) win.close();
-      ipcMain.removeHandler("edit-doctor"); // remove existing handler
-    })
-
-    /* close form when the renderer process informs that the edit process is finished */
-    ipcMain.on("doctor-form-finish", (event, args) => {
-      if(win) win.close();
-      parentWindow.webContents.send("reload-data");
+      console.log(container.childElementCount);
+      workCount--;
+    
     });
   }
 
+
+
+}
+
+
+function toggleInputs (method) {
+  const inputs = document.querySelectorAll("input");
+
+  inputs.forEach(
+    input => {
+      if (input.getAttribute("id") != "id" && input.getAttribute("id") != "doctorId"
+          && input.getAttribute("id") != "specialization") {
+        if (method === "PUT")
+          input.removeAttribute("readonly");
+        else
+          input.setAttribute("readonly", true);
+
+      }
+    }
+  );
+
+  if (method === "PUT") {
+    const editButton = document.getElementById("edit-doctor");
+    editButton.style.display = "block";
+    const deleteButton = document.getElementById("delete-doctor");
+    deleteButton.style.display = "block";
+  }
+  else {
+    const editButton = document.getElementById("edit-doctor");
+    editButton.style.display = "none";
+    const deleteButton = document.getElementById("delete-doctor");
+    deleteButton.style.display = "none";
+
+  }
+}
+
+/* To convert the input data to yyyy-mm-dd format*/
+function formatDate(input) {
+    console.log(input);
+    var d = new Date(input),
+        month = '' + (d.getMonth() + 1),
+        day = '' + d.getDate(),
+        year = d.getFullYear();
+
+    if (month.length < 2)
+        month = '0' + month;
+    if (day.length < 2)
+        day = '0' + day;
+
+    return [year, month, day].join('-');
+}
+
+
+
+
+/* Show error message */
+function showErrorMessage(message) {
+
+  // clear any error messags
+  while (errorDiv.lastChild)
+    errorDiv.removeChild(errorDiv.lastChild);
+
+  let errorNode = document.createElement('div');
+  errorNode.setAttribute('class', 'alert alert-danger');
+  errorNode.setAttribute('role', 'alert');
+  errorNode.innerHTML = message;
+  errorDiv.appendChild(errorNode);
+}
+
+/* get doctors data by id*/
+async function getDoctorById (id) {
+  try {
+    const response = await fetch(`${serverUrl}/api/doctors/${id}`, {
+      method: "GET",
+      headers: {
+        "Content-Type" : "application/json",
+        "Accept" : "application/json"
+      }
+    });
+
+    return response;
+  }
+  catch (error) {
+
+    console.error(error);
+  }
+}
+
+/* edit doctor data by id*/
+async function editDoctorById (id, data) {
+  try {
+    const response = await fetch(`${serverUrl}/api/doctors/${id}`, {
+      method: "PUT",
+      headers: {
+        "Content-Type" : "application/json",
+        "Accept" : "application/json"
+      },
+      body: JSON.stringify(data)
+
+    });
+
+    return response;
+  }
+  catch (error) {
+    console.error(error);
+  }
+}
+
+/* delete doctor data by id */
+async function deleteDoctorById (id){
+  try {
+    const response = await fetch(`${serverUrl}/api/doctors/${id}`, {
+      method: "DELETE",
+      headers: {
+        "Content-Type" : "application/json",
+        "Accept" : "application/json"
+      }
+    });
+
+    return response;
+  }
+  catch (error) {
+    console.error(error);
+  }
 }
